requirements.txt
cryptography
opencv-python
pillow